Extract Repository::InstanceMethods to a separate file, move tests
